create src/parsers/ to home xml-feed parser (renamed from feed-parser), web-scraper parsers, and textParts parser
