+++ src/parsers/text/__tests__/textPart.test.node.ts
