+++ src/parsers/xml-feed/__tests__/fixtures/getFixtures.ts
@@ -1,4 +1,4 @@
-import { readdirSync, readFileSync } from 'fs';
+import { readdirSync, readFileSync } from 'node:fs';
 import type { RssFeed } from '../../parser';
 import { getRssParser } from '../../parser';
