+++ src/parsers/xml-feed/__tests__/fixtures/invalid/deal-missing-description.xml
