+++ src/parsers/xml-feed/__tests__/fixtures/regenValidFixtures.ts
