+++ src/parsers/xml-feed/__tests__/fixtures/valid/ozbargain-rss-small.parsed.json
