+++ src/parsers/xml-feed/__tests__/local.ts
@@ -1,4 +1,4 @@
-import { readFile } from 'fs/promises';
+import { readFile } from 'node:fs/promises';
 import type { OzbargainFeed } from '../parser';
 import { convertToOzbargainFeed } from '../parser';
 import { parseRssFeedFromString } from './fixtures/getFixtures';
