+++ src/screens/dealsFeed/dealCard/renderTitle.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { Text } from '../../../base/components/text/text';
-import type { PartedText } from '../../../feed-parser/textParts';
+import type { PartedText } from '../../../parsers/text/textParts';
 
 export function renderDealCardTitle(text: PartedText): React.JSX.Element {
   return (
